added kabsch alignment

+++ openfold/utils/multi_chain_permutation.py
@@ -50,11 +50,11 @@ def kabsch_rotation(P: torch.Tensor, Q: torch.Tensor) -> torch.Tensor:
 
     # Firstly, compute SVD of P.T * Q
     u, _, vt = torch.linalg.svd(torch.matmul(P.to(torch.float32).T,
-                                             Q.to(torch.float32)))
+                                             Q.to(torch.float32)).to(torch.float32))
     # Then construct s matrix
     s = torch.eye(P.shape[1], device=P.device)
     # correct the rotation matrix to ensure a right-handed coordinate
-    s[-1, -1] = torch.sign(torch.linalg.det(torch.matmul(u, vt)))
+    s[-1, -1] = torch.sign(torch.linalg.det(torch.matmul(u, vt).to(torch.float32)))
     # finally compute the rotation matrix
     r_opt = torch.matmul(torch.matmul(u, s), vt)
     assert r_opt.shape == torch.Size([3,3])
